refactor: change file name of application-form.ts
카멜 케이스 적용

(cherry picked from commit a4299edd5968c5474f8bf4afff5a821911e70e24)
