+++ frontend/src/api/index.ts
@@ -1,3 +1,3 @@
 export * from "./user";
 export * from "./recruitments";
-export * from "./application-forms";
+export * from "./applicationForms";
